Beginning jumping return, attaching fields.
At the moment the jumping returns just throw the value. Obviously this
needs to tag which method it is returning for, and then be caught at the
top of that method.

Fields declared readable and writable now have getters and setters
attached. This does not yet take into account publicity.