Merging all native functionality into $.
The gracelib is a bit of a mess, and nothing works right now, but the
resulting compiled source is a lot tidier because primitive values don't
need to be wrapped in order for them to be used as Grace values.

Method definition, calling, object creation, and vararg definition are
now all performed through the $ function, which switches over the types
of its arguments to perform the correct operation.