Lexical returns and proper vararg types.
The use of `this` meant that the method return wasn't lexically bound,
and stopped working inside of local blocks. This is now saved into a
`$return` variable instead. A more elegant solution might be preferable
in the future.

Varargs was leaving out the type of the varargs themselves. This is now
added to the output.