General improvements of Javascript interop.
Object methods are now no longer actually attached to objects, but are
instead applied if no other method is found on the object. This means
that they will exist on native Javascript objects as well (the one
exception to this is `done`, which is represented as undefined).

Method calls now also allow for getters and setters on native Javascript
objects. Setters will work for any name, and getters will exist for any
property that does not already contain a function, including those that
aren't defined (these return `nothing`, but that isn't fully implemented
yet). Note that it's impossible to retrieve a block from a getter,
because this call system will assume it is a method, and invoke it
straight away.